Reorganized authorization related components and utilities to resolve warnings being reported by linting rules.
